t/xml/notations-1.dat

#data
<!DOCTYPE a[
<!N>
]><a/>
#errors
2;1;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!N
#errors
2;1;bogus comment;;;m
2;3;unclosed internal subset;;;m
2;3;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NO>
]><a/>
#errors
2;1;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NO
#errors
2;1;bogus comment;;;m
2;4;unclosed internal subset;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOT>
]><a/>
#errors
2;1;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOT
#errors
2;1;bogus comment;;;m
2;5;unclosed internal subset;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIO>
]><a/>
#errors
2;1;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOTATIO
#errors
2;1;bogus comment;;;m
2;9;unclosed internal subset;;;m
2;9;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION>
]><a/>
#errors
2;11;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION
#errors
2;10;unclosed md;;;m
2;10;unclosed internal subset;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIONS>
]><a/>
#errors
2;11;no space before md name;;;m
2;12;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION S "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATIONS
#errors
2;11;no space before md name;;;m
2;11;unclosed md;;;m
2;11;unclosed internal subset;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION b>
]><a/>
#errors
2;13;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b
#errors
2;12;unclosed md;;;m
2;12;unclosed internal subset;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION b >
]><a/>
#errors
2;14;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "">
| <{} a>

#data escaped
<!DOCTYPE a[
<!NOTATION b\u0020
#errors
2;13;unclosed md;;;m
2;13;unclosed internal subset;;;m
2;13;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "">

#data
<!DOCTYPE a[
<!NOTATION bcd>
]><a/>
#errors
2;15;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd
#errors
2;14;unclosed md;;;m
2;14;unclosed internal subset;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc  #
#errors
2;17;string after DOCTYPE name;;;m
2;17;unclosed internal subset;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION abc "" "">

#data escaped
<!DOCTYPE a[
<!NOTATION abc PUBLIC\u0020
#errors
2;22;unclosed md;;;m
2;22;unclosed internal subset;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION abc "" "">

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC #
#errors
2;23;string after PUBLIC;;;m
2;23;unclosed internal subset;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION abc "" "">

#data escaped
<!DOCTYPE a[
<!NOTATION abc SYSTEM\u0020
#errors
2;22;unclosed md;;;m
2;22;unclosed internal subset;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION abc "" "">

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC #
#errors
2;23;string after PUBLIC;;;m
2;23;unclosed internal subset;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a"x
#errors
2;26;string after PUBLIC literal;;;m
2;26;unclosed internal subset;;;m
2;26;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" x
#errors
2;27;string after PUBLIC literal;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM "a"x
#errors
2;26;string after SYSTEM literal;;;m
2;26;unclosed internal subset;;;m
2;26;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM "a" x
#errors
2;27;string after SYSTEM literal;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" "b"x
#errors
2;30;string after SYSTEM literal;;;m
2;30;unclosed internal subset;;;m
2;30;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" "b" x
#errors
2;31;string after SYSTEM literal;;;m
2;31;unclosed internal subset;;;m
2;31;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC>
]><a/>
#errors
2;22;no PUBLIC literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC >
]><a/>
#errors
2;23;no PUBLIC literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"
#errors
2;27;unclosed md;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "">

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg" "hij">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg""hij">
]><a/>
#errors
2;28;no space before system literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg'"hij">
]><a/>
#errors
2;28;no space before system literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"'hij'>
]><a/>
#errors
2;28;no space before system literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg''hij'>
]><a/>
#errors
2;28;no space before system literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg" "hij"
#errors
2;33;unclosed md;;;m
2;33;unclosed internal subset;;;m
2;33;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM>
]><a/>
#errors
2;22;no SYSTEM literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM >
]><a/>
#errors
2;23;no SYSTEM literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "efg">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"
#errors
2;27;unclosed md;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "efg">

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"x>
]><a/>
#errors
2;28;string after PUBLIC literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"x
#errors
2;28;string after PUBLIC literal;;;m
2;28;unclosed internal subset;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg'x
#errors
2;28;string after PUBLIC literal;;;m
2;28;unclosed internal subset;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"x>
]><a/>
#errors
2;28;string after SYSTEM literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "efg">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"x
#errors
2;28;string after SYSTEM literal;;;m
2;28;unclosed internal subset;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM 'efg'x
#errors
2;28;string after SYSTEM literal;;;m
2;28;unclosed internal subset;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM 'efg'xyz<!
#errors
2;28;string after SYSTEM literal;;;m
2;32;unclosed internal subset;;;m
2;32;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIOn b SYSTEM "x">
]><a/>
#errors
2;3;lowercase keyword;NOTATION;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!notation b SYSTEM "x">
]><a/>
#errors
2;3;lowercase keyword;NOTATION;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!nOTATION b SYSTEM "x">
]><a/>
#errors
2;3;lowercase keyword;NOTATION;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b SYSTEM "x">
<!NOTATION b SYSTEM "y">
]><a/>
#errors
3;1;duplicate notation decl;;b;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b SYSTEM "x">
%z;
<!NOTATION c SYSTEM "y">
]><a/>
#errors
3;1;stop processing;;;i
3;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
|   <!NOTATION c "" "y">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "ab" 'efg'xy
#errors
2;33;string after SYSTEM literal;;;m
2;34;unclosed internal subset;;;m
2;34;no root element;;;m
#document
| <!DOCTYPE a>
